fix(skippy): clear PR clippy findings
